+++ tests/unit/processor/string_splitter/test_string_splitter.py
@@ -1,4 +1,6 @@
 # pylint: disable=missing-docstring
+# pylint: disable=too-many-arguments
+# pylint: disable=too-many-positional-arguments
 import re
 
 import pytest
@@ -6,52 +8,146 @@ import pytest
 from tests.unit.processor.base import BaseProcessorTestCase
 
 test_cases = [
-    (
-        "splits without delimeter on whitespace",
+    pytest.param(
         {
             "filter": "message",
-            "string_splitter": {"source_fields": ["message"], "target_field": "result"},
+            "string_splitter": {
+                "source_fields": ["message"],
+                "target_field": "result",
+                "drop_empty": True,
+            },
         },
         {"message": "this is the message"},
-        {"message": "this is the message", "result": ["this", "is", "the", "message"]},
+        ["this", "is", "the", "message"],
+        id="splits_without_explicit_set_delimiter_on_whitespace",
     ),
-    (
-        "splits with delimeter",
+    pytest.param(
         {
             "filter": "message",
             "string_splitter": {
                 "source_fields": ["message"],
                 "target_field": "result",
                 "delimiter": ", ",
+                "drop_empty": True,
             },
         },
         {"message": "this, is, the, message"},
-        {"message": "this, is, the, message", "result": ["this", "is", "the", "message"]},
+        ["this", "is", "the", "message"],
+        id="splits_with_delimiter",
     ),
-]  # testcase, rule, event, expected
-
-failure_test_cases = [
-    (
-        "splits without delimeter on whitespace",
+    pytest.param(
         {
             "filter": "message",
-            "string_splitter": {"source_fields": ["message"], "target_field": "result"},
+            "string_splitter": {
+                "source_fields": ["message"],
+                "target_field": "result",
+                "delimiter": ",",
+                "drop_empty": True,
+            },
         },
-        {"message": ["this", "is", "the", "message"]},
-        {"message": ["this", "is", "the", "message"], "tags": ["_string_splitter_failure"]},
-        ".*ProcessingWarning.*",
+        {"message": "this,"},
+        ["this"],
+        id="splits_one_item_with_delimiter",
     ),
-    (
-        "splits without delimeter on whitespace",
+    pytest.param(
         {
             "filter": "message",
-            "string_splitter": {"source_fields": ["message"], "target_field": "message"},
+            "string_splitter": {
+                "source_fields": ["message"],
+                "target_field": "result",
+                "delimiter": ",",
+                "drop_empty": True,
+            },
         },
-        {"message": "this is the message"},
-        {"message": "this is the message", "tags": ["_string_splitter_failure"]},
-        ".*FieldExistsWarning.*",
+        {"message": ",,this,,"},
+        ["this"],
+        id="splits_one_item_with_multiple_delimiter_and_drop_empty",
     ),
-]  # testcase, rule, event, expected, error_message
+    pytest.param(
+        {
+            "filter": "message",
+            "string_splitter": {
+                "source_fields": ["message"],
+                "target_field": "result",
+                "delimiter": ",",
+                "drop_empty": False,
+            },
+        },
+        {"message": ",,this,,"},
+        ["", "", "this", "", ""],
+        id="splits_one_item_with_multiple_delimiter_and_no_drop_empty",
+    ),
+    pytest.param(
+        {
+            "filter": "message",
+            "string_splitter": {
+                "source_fields": ["message"],
+                "target_field": "result",
+                "delimiter": ",",
+                "drop_empty": True,
+            },
+        },
+        {"message": " , ,this, ,"},
+        ["this"],
+        id="splits_one_item_with_multiple_delimiter_and_empty_fields",
+    ),
+    pytest.param(
+        {
+            "filter": "message",
+            "string_splitter": {
+                "source_fields": ["message"],
+                "target_field": "result",
+                "delimiter": ",",
+                "drop_empty": True,
+            },
+        },
+        {"message": ",, this , , "},
+        [" this "],
+        id="splits_one_item_with_multiple_delimiter_and_whitespace",
+    ),
+    pytest.param(
+        {
+            "filter": "message",
+            "string_splitter": {
+                "source_fields": ["message"],
+                "target_field": "result",
+                "delimiter": ",",
+                "drop_empty": True,
+            },
+        },
+        {"message": "\n,,this,\t, "},
+        ["this"],
+        id="splits_one_item_with_multiple_delimiter_and_newline",
+    ),
+    pytest.param(
+        {
+            "filter": "message",
+            "string_splitter": {
+                "source_fields": ["message"],
+                "target_field": "result",
+                "delimiter": ",",
+                "drop_empty": True,
+            },
+        },
+        {"message": ",, this, , "},
+        [" this"],
+        id="splits_one_item_with_multiple_delimiter_and_whitespace_only_in_front",
+    ),
+    pytest.param(
+        {
+            "filter": "message",
+            "string_splitter": {
+                "source_fields": ["message"],
+                "target_field": "result",
+                "delimiter": ",",
+                "drop_empty": True,
+            },
+        },
+        {"message": "hello , world,this, is a very complex,\n , and even multiline, text,,, "},
+        ["hello ", " world", "this", " is a very complex", " and even multiline", " text"],
+        id="splits_one_item_with_multiple_delimiter_and_whitespace_only_in_front",
+    ),
+]
 
 
 class TestStringSplitter(BaseProcessorTestCase):
@@ -60,16 +156,40 @@ class TestStringSplitter(BaseProcessorTestCase):
         "rules": ["tests/testdata/unit/string_splitter/rules"],
     }
 
-    @pytest.mark.parametrize("testcase, rule, event, expected", test_cases)
-    def test_testcases(self, testcase, rule, event, expected):  # pylint: disable=unused-argument
+    @pytest.mark.parametrize(["rule", "event", "expected"], test_cases)
+    def test_testcases(self, rule, event, expected):
         self._load_rule(rule)
         self.object.process(event)
-        assert event == expected
+        assert event["result"] == expected
 
-    @pytest.mark.parametrize("testcase, rule, event, expected, error_message", failure_test_cases)
-    def test_testcases_failure_handling(self, testcase, rule, event, expected, error_message):
+    @pytest.mark.parametrize(
+        ["rule", "event", "expected", "error_message"],
+        [
+            pytest.param(
+                {
+                    "filter": "message",
+                    "string_splitter": {"source_fields": ["message"], "target_field": "result"},
+                },
+                {"message": ["this", "is", "the", "message"]},
+                {"message": ["this", "is", "the", "message"], "tags": ["_string_splitter_failure"]},
+                ".*ProcessingWarning.*",
+                id="splits_without_delimiter_on_whitespace_with_no_string",
+            ),
+            pytest.param(
+                {
+                    "filter": "message",
+                    "string_splitter": {"source_fields": ["message"], "target_field": "message"},
+                },
+                {"message": "this is the message"},
+                {"message": "this is the message", "tags": ["_string_splitter_failure"]},
+                ".*FieldExistsWarning.*",
+                id="splits_without_delimiter_on_whitespace_with_existing_field",
+            ),
+        ],
+    )
+    def test_testcases_failure_handling(self, rule, event, expected, error_message):
         self._load_rule(rule)
         result = self.object.process(event)
         assert len(result.warnings) == 1
         assert re.match(error_message, str(result.warnings[0]))
-        assert event == expected, testcase
+        assert event == expected